Properly share trampolines between instances

+++ src/koffi/src/call.hh
@@ -41,7 +41,7 @@ class alignas(8) CallData {
     Span<uint8_t> old_stack_mem;
     Span<uint8_t> old_heap_mem;
 
-    int16_t used_trampolines = 0;
+    uint32_t used_trampolines = 0;
 
     LocalArray<OutArgument, MaxOutParameters> out_arguments;
